+++ docs/history.rst
@@ -111,3 +111,12 @@ History
 ----------------------
 
 * Fixed bug with ``ampScale`` argument in ``cw_pulse_generator()``.
+
+
+2021.05.1 (2021-05-14)
+----------------------
+
+* Added N5194A Vector UXG support for PDW format 3 rev B that was introduced in firmware A.01.30.  Updated vector uxg example names to better reflect what the examples do.
+* Refined N5193A file streaming to define size of PDW block
+* Reformatted comments related to analog and vector PDW sections to be more easy to read.
+